LC-14 - code review changes.
