+++ src/main/java/uk/gov/cshr/civilservant/dto/factory/CivilServantDtoFactory.java
@@ -2,13 +2,15 @@ package uk.gov.cshr.civilservant.dto.factory;
 
 import org.springframework.stereotype.Component;
 import uk.gov.cshr.civilservant.domain.CivilServant;
+import uk.gov.cshr.civilservant.domain.OrganisationalUnit;
 import uk.gov.cshr.civilservant.domain.Profession;
 import uk.gov.cshr.civilservant.dto.CivilServantDto;
+import uk.gov.cshr.civilservant.dto.OrganisationalUnitDto;
 
 import java.util.stream.Collectors;
 
 @Component
-public class CivilServantDtoFactory {
+public class CivilServantDtoFactory extends DtoFactory<CivilServantDto, CivilServant>  {
 
     public CivilServantDtoFactory() {
